requirements.txt
Flask
python-dotenv
cohere
